// src/Utils/types.ts
export type IDType = string | number;

export interface ImageType {
  src?: any;
  alt?: string;
  width?: number;
  height?: number;
  loading?: "lazy" | "eager";
}
// ...
export interface ReviewType {
  average?: number;
  count?: number;
  rating_details?: {
    "5": number;
    "4": number;
    "3": number;
    "2": number;
    "1": number;
  };
  items?: Array<{
    id?: IDType;
    user?: {
      name?: string;
      avatar?: ImageType;
    };
    rating?: number;
    title?: string;
    review?: string;
  }>;
}
// ...
export interface ICourse {
  cc?: string[];
  title?: string;
  slug?: string;
  path?: string;
  published_at?: string;
  category?: string;
  thumbnail?: any;
  price?: number;
  currency?: string;
  total_lectures?: any;
  total_students?: number;
  language?: string;
  instructor?: IDType;
  duration?: string;
  membership?: string[];
  isPopular?: boolean;
  excerpt?: string;
  description?: any;
  reviews?: ReviewType;
  curriculum?: IDType[];
  teacher?: any;
  id?: any;
  fullname?: any;
  image_url?: any;
  course?: any;
  pathBuy?: string;
  withoutDiscountPrice?: any;
  installmentPaymentPrice?: any;
  section?: any;
}

export interface BlogMetaType {
  title?: string;
  slug?: string;
  path?: string;
  tags?: any;
}

export interface IBlog {
  title?: string;
  slug?: string;
  path?: string;
  postedAt?: string;
  image?: ImageType;
  category?: BlogMetaType;
  tags?: BlogMetaType[];
  views?: number;
  author?: IInstructor;
  content?: string;
  excerpt?: string;
}

export interface ISocial {
  label?: string;
  icon?: string;
  url?: string;
}

export interface IInstructor {
  id?: IDType;
  name?: string;
  slug?: string;
  path?: string;
  url?: string;
  image?: ImageType;
  designation?: string;
  bio?: string;
  socials?: ISocial[];
}